+++ example/shutdown.dart
@@ -1,17 +1,18 @@
 import 'dart:async';
 
-import 'package:shutdown/shutdown.dart';
+import 'package:shutdown/shutdown.dart' as shutdown;
 
 Future main() async {
-  registerDefaultProcessSignals();
+  shutdown.triggerOnSigInt();
+  shutdown.triggerOnSigHup();
 
   final db = await _acquireDB();
-  addShutdownHandler(() => db.close());
+  shutdown.addHandler(() => db.close());
 
   // [do you stuff]
 
   // call it at the end, this is a successful exit
-  await shutdown();
+  await shutdown.shutdown();
 }
 
 Future _acquireDB() async => null; // TODO: implement